Fix inconsistent real path canonicalization in SdfLayer
Centralize call to Sdf_CanonicalizeRealPath to ensure
layer real paths are always in a consistent format.
In particular, canonicalization was not happening
when SdfLayer::UpdateAssetInfo was called. On Windows,
calling this function would cause the layer's real
path's formatting to change and notifications to
be sent.

This fix causes SdfLayer to pass the canonicalized
path to SdfFileFormat::Read when SdfLayer::Reload
is called. TestPcpDynamicFileFormatPlugin uses this
path to author references and payloads; it now
normalizes the path to ensure a consistent format
for baseline comparisons.

Note that SdfLayer is currently inconsistent with the
paths it passes to SdfFileFormat. Sometimes it's the
canonicalized path, sometimes it's not. This is
something that will improve with Ar 2.0, since Sdf
will no longer be in the business of canonicalizing
paths. So, I've chosen not to fix this inconsistency
here.

This fix allows us to remove the workaround in
testUsdExternalAssetDependencies.

(Internal change: 2129519)