[ar 2.0] Make ArResolver::GetModificationTimestamp optional
ArResolver now provides a default implementation for
ArResolver::GetModificationTimestamp that just returns an
invalid ArTimestamp.

SdfLayer no longer requires a layer to have a valid timestamp.
Previously SdfLayer would issue a coding error and fail if a
timestamp could not be retrieved, even if the layer itself was
opened and read successfully.

The timestamp was used by SdfLayer::Reload to determine if
a layer had been changed externally and needed to be reloaded.
SdfLayer::Reload will now consider any layer with an invalid
timestamp and any layer with external asset dependencies with
invalid timestamps as "dirtied" and will always reload that
layer.

(Internal change: 2189596)
(Internal change: 2189952)