Fix code sites where layer identifiers with file format arguments were being passed down to ArResolver. File format arguments are an Sdf-level concern that resolver plugins should not have to reason about.
These code sites have been updated to extract the layer path
from the identifier and pass that down to ArResolver instead.
In some cases, the layer path was already available so no
additional work was needed.

In the future, we may want to look at storing a layer's
identifier, asset path, and args separately and using
different types than just std::string to get the compiler
to help us avoid passing in the full identifier to Ar.
That will be easier to do once the deprecated Ar 1.0 code
is removed. In the meantime, splitting the identifier as
needed is the simplest fix.

(Internal change: 2196373)
(Internal change: 2196470)