Fix crashes when reading .usda files containing certain strings with escaped quotes.
Strings like "\" would be accepted by the parser but then
cause Sdf_EvalQuotedString to crash because no closing
quote would be found. This change fixes the parser to
reject such strings. As a safeguard, this change also
modifies Sdf_EvalQuotedString to avoid reading off the
end of the given string.

This fix exposed a few test cases that had previously
been accepted by the parser but should not have been,
like:

string s = "fo\\"o"

THe parser now rejects these strings, since the double
"\\" is an escaped backslash and does not apply to the
subsequent quote. These test cases were updated to
escape the quote as well.

Fixes #1630

(Internal change: 2242772)