Add support for "detached" layers in Sdf
A detached layer is detached from its serialized backing
store, isolating it from any external changes to the
serialized data. Prior to this change, whether a layer was
detached was determined solely by the layer's file format
plugin and the ArAsset implementation (if one was used by the
file format).

For example, .usdc layers are typically *not* detached layers.
The Crate file format with the default ArFilesystemAsset holds
on to a memory mapping for the file on disk and reads data from
it as needed. In this case, having a .usdc layer open in a
process and overwriting the file in another can lead to crashes
or other workflow issues. The new detached layer feature lets
users to specify layers that should not maintain persistent
connections to their data store. In the example above, a
detached .usdc layer would not maintain a memory mapping to
the source file on disk. This would allow other processes to
overwrite the file without introducing the issues described
above.

Detached layers are specified via include/exclude patterns
in the new SdfLayer::DetachedLayerRules object. SdfLayer
will do simple substring matches against these patterns. For
example, a user could specify that all layers should be
detached, except for those that contain "Sim" or "FX" in
their identifier.

New virtual methods for initializing and reading to a detached
layer were added to the SdfFileFormat interface to support this
feature. The default implementations use an SdfData object, which
is the same in-memory data container used by the .usda file format.

A new virtual method was also added to SdfAbstractData to query
whether that data object is detached. By default, this returns
true if the data object is not streaming, false otherwise.

The default implementation for SdfFileFormat::_ReadDetached will
issue a warning if it detects that the file format produced a
non-detached data object from Read and perform a (potentially
expensive) copy of the layer data into a new SdfData object.
File formats that are able to provide a more efficient
implementation should do so, but for convenience they can
opt in to the copying behavior by just calling
_ReadAndCopyLayerDataToMemory in their implementation.

The main takeaway is that non-streaming file formats typically
will not have to implement anything new, since they are already
classified as detached by default. Streaming file formats will
need to implement at least _ReadDetached.

Fixes #1852

(Internal change: 2247699)