sdf/pcp: Prohibit variant selections in composition arc paths
This is now enforced both at the Sdf schema level and at the
Pcp composition level.  Schema enforcement means that
invalid paths cannot be authored via runtime API, nor will
they be accepted by the text parser.

As a staging measure, the environment setting
SDF_SCHEMA_PROHIBIT_INVALID_VARIANT_SELECTIONS
may be set to '0' to continue to allow these paths at the
schema & parser level.  Regardless of this setting, the
composition system will always prohibit such paths where
invalid for composition purposes.

Background

SdfPath is a namespace location value type used for several
related purposes.  A main purpose is to identify objects
(UsdObject) in the scene namespace.  Supporting this,
SdfPath is also used to address backing scene description
storage (SdfSpec, aka "specs") in an SdfLayer.  SdfPath
sees further use in Hydra to identify entries in the
scene index.

Given that one type is shared across these subsystems,
some attention is required as to which subset of path
capabilities are meaningful in each.

Variants are a composition system (Pcp) concept that is driven
by interpreting scene description.  Storage addressing for
layer opinions for variants uses paths with embedded variant
selections, for example </A{x=y}B>, to specify the branch of
spec storage corresponding to the variant.  Such paths, however,
are not meaningful to the composed scenegraph, and are not valid
in composition arcs.

In this bug, there were scenarios where Pcp was attempting
to interpret SdfPath values containing variant selections
in composition arc paths.  This change introduces stricter checks
at the Sdf schema level to guard against variant selections in
fields where they are not allowed, as well as stricter checks
and better diagnostics in the Pcp composition system.

(Internal change: 2263171)
(Internal change: 2263363)