sdf: Remove testSdfParsingOld
This test was a maintenance burden as it duplicated the
majority of the test cases in the primary testSdfParsing
test, with only a handful of cases differing due to
their use of "old" attribute typenames. This forced
developers to copy those new tests to multiple locations.

A selection of these old typename tests have been
migrated to testSdfParsing, allowing us to maintain
roughly the same level of test coverage while letting
us get rid of the test duplication.

These old typenames were only used for Pixar-internal
assets prior to USD's open source release in 2016 and
should not be present in any external assets. So,
practically speaking a reduced level of test coverage
shouldn't really matter. We intend to remove support for
reading the old typenames in the future, but in the
interim this change will make it easier for devs to add
test cases when needed.

(Internal change: 2292197)