Allow playback from SdfChangeList.
Currently SdfChangeList entries are ambiguous when it comes to
reproducing the edits they indicate.  For example, an entry can
have both didAddNonInertPrim and didRemoveNonInertPrim but it
can't tell you if it was an add followed by a remove or a remove
followed by an add or even add/remove/add, etc.

With this change entries that would be ambiguous when combined are
kept separate.  There is no change to the API.

(Internal change: 2313470)