NVIDIA: Fix SdfCopySpec issues about overwriting field SdfFieldKeys->TargetPaths/ConnectionPaths
The whole process of SdfCopySpec can be summarized as the following steps:

* Getting all fields to be copied, including value fields and children fields.
* Copying value fields firstly.
* Then copying children fields.

It's possible that the children fields fetched in step 1 is cleared in step 2, so step 3 is failed with copying a non-existent children field. This fix is to delay the fetch of children fields after step 2 so step 3 will be done on the correct children fields.

Fixes #3095
(github pull request #3117)

- PR updated internally to enhance SdfCopySpec() to handle overlapping source/destination by copying the source to a temporary anonymous layer first, then copying that temporary to the destination.

(Internal change: 2335761)