Verify SdfSpecType is valid for GetSpecDefinition.
A crash was discovered which was caused by a memory access violation within SdfSchemaBase::GetSpecDefinition.

The root of the problem is that SdfSchemaBase::_specDefinitions is indexed based on SdfSpecType which is just an enum. A USDC file can be maliciously crafted to force a SdfSpecType to be any int value which will be implicitly converted to an SdfSpecType enum. This invalid SdfSpecType will be passed to SdfSchemaBase::GetSpecDefinition triggering the crash.

SdfSchemaBase::GetSpecDefinition is public and does not provide any checks that the input SdfSpecType is a valid value. A check that the input SdfSpecType is in the proper range of SdfSchemaBase::_specDefinitions has been added to SdfSchemaBase::GetSpecDefinition

(Internal change: 2342694)
(Internal change: 2342712)