Updating Pcp change processing to correctly handle changes to the defaultPrim layer metadata when the field contains a path instead of just a root prim name. Previously, the change processing code assumed the defaultPrim token was a root prim name so path values would end up treated as an empty default prim causing a resync of everything depending on the layer instead of the just the prims that depend on the default prim path. This changes fixes that. Additionally, I made an optimization to the change processing that no longer invalidates dependencies on the new default prim path, just dependencies on the old default prim path.
I've also also a couple static helpers to SdfLayer to convert between
tokens and paths for the default prim field. The Pcp change management
and namespace editing code have been updated to use these functions.

Lastly, I'll note that the change processing for default prim changes is
still overzealous in that it can cause resyncs on any prim indexes that
depend on the old path, not just prim indexes that depend on the
default prim of the layer. I left an XXX comment expressing this in
PcpChanges and will file a bug.

(Internal change: 2361551)