sdf: Mark layers as dirty on identifier change
Changing a layer's identifier may update its resolved path to
point to a different file whose contents may differ from what's
currently in the layer. In that case, we want to ensure that
calling Save() on that layer will write its contents out to
that file. However, Save() skips writing if the layer is not
dirty and the file already exists.

This change causes layers to be marked dirty any time their
resolved paths change due to an identifier change, which avoids
the issue above. Note that this is conservative; this will
also cause Save() to write a layer out even if its contents
are the same as the file at the new resolved path. This is
the best we can do, since checking whether the serialized data
actually matches the layer would be too	expensive.

Fixes #3386

(Internal change: 2371658)